+++ 04.class/memo-manager.js
